Task: Update styles based on old theme.
The original ontario theme used green, black and white
as the base colors. I've added this back in with
minimal changes to the style sheet. As with the
previous theme, work is still needed.

Also added some styling to the topics for smaller screens
as they were hard to read when rendering on small screens
due to how the background changes by default in CKAN themes.

Added background image to home page and removed textured
background images. The background image used is a CC0
license.