Improvement: Form should display saved values.
Non-input form elements were not displaying previously saved
values because selected was hard coded values.
This was updated so that the selected value would be set
based on the data passed into the form.

Also updated the display format of these values to be more
human readable on the dataset page.