Task: Add test coverage for validations.
Test each validation function with it's own class in test_plugin.py.
Different functions / methods were used for each test scenario within classes.

Remove instance checking for dates. I'm only saving string representations of dates.
This was also causing some odd behaviour. In python it would pass
for datetime.datetime as well as datetime.date objects but when using
the api I can only pass in string objects of datetime type which in python and
ruby raised the exception as they are stringified objects. No need to
convert to datetime objects as I'm only saving the string versions anways
in the extras fields.

Fix bugs in validations caused when trying to lower() values passed in. This crashed
when trying to lower non strings and tests caught this.