Task: Turn stats into smarties.
Turn the default stats block into smarties to follow ontario
style guide for this type of information.

Switch css to less to align with CKAN and also make styling a bit more
modular for the future. This is a basic switch at this point and does not
use a formal build system at this time.

Broke new smarty styling into own less file and compile to css and merge
withthe ontario_theme.css which is added as a resource.

Change layout to have single column rows for stats and activity feed.

Fixed up styling for skittles (topics) to match style guide instead of home page
which doesn't follow the style guide exactly.

Had to play with styling to prevent conflicts with main.css and the new
theme styling.