Improvement: Expand test coverage.
Add test coverage for create and update actions of
datasets. This test coverage focuses on the extrafields
that I've added custom or more detailed validations rules to
in the schemas.

Update exemption validation so that empty values are not allowed.
Instead these are modified to 'none'. For now I'm still ignoring missing
keys.