Bug: node_id validation was being triggered even when missing.
The is_positive_integer validator was always triggering even
when the value was empty. For some reason it seemes that
ignore_missing wasn't stopping validation as it was supposed to.
This validation does clal int_validator which returns None for empty
values however, this should not have been called if ignore_missing
was being called first properly.

This validation gives the desired result but does allow negative numbers
which shouldn't be an issue.