Bug: When loading org page throws error.
When user went to an organization page a server error was shown.
This was due to an AttributeError being raised because IFacets couldn't
find the the group_facets() and organization_facets() functions.

CKAN doesn't check to see if the functions are implemented, only
that IFacet is implemented in an extension and then calles the
extensions.*_facets() function. So these must be set even if not modifying
the facet (e.g. `pass`). In my case I am fine with modifing the facets here as well.