wip: Filter datasets by those that have resources.
Add sort option to pull datasets with resources to the top of the search results.
This sorts by greatest number of resources attached to a dataset at the top.

Implement IPackageController to alter search params before_search(). This is
needed as extra query strings are added to the solr filter (fq) but the value is
wrapped in double quotes and returns the wrong result. Updating the value
to strip these quotes.

Quick implementation of a filter in the secondary content.