Task: Tests, refactor and add remove filter link.
Add tests for fq filter values.

Add remove url for when the new filter is set.

Refactor plugin.py to bring function up to make it a little
easier to test. Also, had to modify how I was removing double
quotes from value. I was originally doing it on any double quotes
in the value but realized the value could have multiple filters applied.
I changed it to do an exact match. This means it only works for the exact
range I've set which is alright because I can still access 0 datasets (the opposite
of what I've set) by using ?num_resources=0.