Task: Customize resource uploader with IUploader interface.
Implement IUploader to only allow resource uploads based on a
whitelist of resource mimetypes.

This overrides the __init__ function of the ResourceUpload class
to validate resource file type using magic. Modifications are marked
with start and end comments.

This could possibly be moved to a validator instead of this implementation.

Accepted resources are based on the default resource_formats.json
list from ckan but with exe and html for now.