Task: Cleanup header template and lang toggle.
The header_account_logged block was added a few months
ago to change the user dashboard link to go to the users
datastes instead of activity feed. This added a block outside
of its parent block(s) where it we were already calling super().

Language selector did not appear in the non-authed user header.
While adding this I also moved it to a snippet.