Bug: Resource item edit link gave 404.
This is due to using ckan master template as the base
instead of the ckan version we're using (2.8.2). Some
routes are using flask and others are using pylons and
each version more flask views are introduced that replace
the old controllers but this breaks if the migration hasn't
happened yet for a controller.