Task: Replace top tags on homepage with new keywords.
Add new helper to get the keywords facet. The get_facet_items_dict()
that is used by defualt is based on the default facets or those passed into it.
The home.py view doesn't check if an extension adds custom facets. To
get around this the helper does it's own search and exposes it to the templates.

Add mutli-linguagal keyword support to the home page.

Add tests for new helper.