Task: Remove licence_range_start and licence_range_end from schema
Licence_range_start and Licence_range_end fields specific to Colby and shouldn't be included in the public schema. It is used in Colby to indicate that the datasets will be available to the OPS for use for a limited time, but no data on the public catalogue will have data with licences with time-limited conditions.