Bug: Let user click first quarter of search form.
Inherit from parent template where able in the page.html template
and add a new custom block that is used in the search pages for
organizations and datasets. This block helps lift the search form
above the `wrapper` class and into it's own row above the 2 column
layout. This wrapper class was blocking a user from clicking into
the search box or remove filters in the first "colulm" of the layout.