Improvement: Fix resource description overlap and display.
Make resource description (ckan core field) multi-lingual in schema.
This currently adds description-en, description-fr instead of _translated.
This was causing the description to not appear in the resource_item.html
because it was not being saved in description_translated or description field.
It was only being saved in description-en and description-fr oddly.

Restructure styles and resource_item.html to display resource details on
a package page in 2 columns. This fixes the description and button
overlap issue.

Override resource_item.html to make the resource display 2 columns
and prevent overlapping of buttons and text. Complete override
needed as there isn't a block wrapping the <li> element. Template
starts showing the description once schema is fixed.

Override resource_read.html to load multi-lingual descriptions.