wip: Modifications.
[x] Update anti_csrf.py to work with Flask requests.
[x] Update plugin for above
[] Set cookie on flask responses

Intercept is still called but just updates the html with a token.
before_app_request handles validation of token matching.
setting the cookie in the response currently only works for pylons responses, or if a pylons response was first given then going to a flask response. Otherwise the response object doesn't exist yet.

Also have to set streaming=False in pylons_app.py.