Bug: Accessing non-existent data_range_start breaks everything.
Using data_range_start to sort in the resources_list template breaks all datasets that have a dataset of type 'data' but lack a data_range_start on all of the resources. Moreover, resources should be grouped by the data_range_start and data_range_end combo, not just data_range_start. a new data_range is created with defaults, so that all data type resources have a data_range. Then it is safe to group and sort by data_range.