Task: Add link to Open Data Guidebook for exemption field.
In order to provide more guidance in the add dataset form, for the exemption field a link to the open data guidebook was added. In order to enable this so that the link could be opened in a new window (since the link is in a form), allow_html is enabled for help text.