Task: Add new fields for organizations.
Update organization schema for 2 new fields to help
more easily categorize the organization and the datasets
assigned to them (e.g. Provincial Ministry dataset).

Index and these new fields to enable faceting. These new
fields are extras and are not included by default in the package
dict passed into `before_index()`. To get this I'm getting the org
and accessing the values that way. This is done in other core extensions.

Facet the new fields for dataset search page. Easiest way was to break
them out of the facet loop to get the proper choices passed in.